Updated to Primefaces 5.0, added email address support
Registration now requires an email address instead of a username (note:
it does not take steps to confirm the email). Usernames are now out, and
thus accounts will need to be remade. The following emails are on the
"developer" list (and thus will have immediate access to the admin
options when their accounts are created):

sql-tutor@googlegroups.com
jake.cobb@gatech.edu
mayur.naik@gmail.com
sumitg@microsoft.com
wholton@gatech.edu
edwardo@cc.gatech.edu
sham@cc.gatech.edu

"sql-tutor@googlegroups.com" has replaced the "dev" account. The
password to this account is "sqltutordev!". This Google group email is
also pointed to by the "Contact Us" button instead of our individual
emails.

The login menu now has a "Forgot Password?" option which sends an email
with a link to reset their password. All of this is encrypted/salted
within the database.

Access control is implemented. The person who uploaded the schema is the
only one who is able to modify the options, instances, and questions.
All of the current schemas have their owner set to
"sql-tutor@googlegroups.com".

The "Schema Instance" page now displays the database as I mentioned. In
addition, I found a solution to the width problem for the
"Reorder/Delete Question" widget - it looks much nicer now.

I've also upped the font size from 14px to 17 (I left the headers at
20).

Primefaces 5.0 is now in.

+++ src/edu/gatech/sqltutor/beans/SchemaOptionsPageBean.java
@@ -62,6 +62,14 @@ public class SchemaOptionsPageBean implements Serializable {
 	
 	public void submit() {
 		try {
+			boolean hasPermissions = databaseManager.checkSchemaPermissions(userBean.getEmail(), userBean.getSelectedSchema());
+			if(!hasPermissions) {
+				final FacesMessage msg = new FacesMessage(FacesMessage.SEVERITY_INFO,
+						"You do not have permissions for this schema.", "");
+				FacesContext.getCurrentInstance().addMessage(null, msg);
+				return;
+			}
+			
 			if(deleteThisSchema) {
 				databaseManager.deleteSchema(userBean.getSelectedSchema());
 				userBean.setSelectedSchema("company"); //TODO: may need to set a warning that company is the default schema
